+++ legislation_analysis/topic_modeling/topic_modeling.py
@@ -12,7 +12,7 @@ from gensim.models.coherencemodel import CoherenceModel
 from gensim.models.ldamodel import LdaModel
 from scipy.stats import loguniform, randint
 
-from legislation_analysis.topic_modeling.abstract_topic_modeling import (
+from legislation_analysis.topic_modeling.base_topic_modeling import (
     BaseTopicModeling,
 )
 from legislation_analysis.utils.constants import (
